+++ src/neural_network/optimizer/adam.rs
@@ -1,5 +1,5 @@
 use crate::neural_network::{Layer, Optimizer};
-use ndarray::{Array2, Array4};
+use ndarray::{Array2, Array3, Array4};
 
 /// Adam optimizer implementation.
 ///
@@ -238,6 +238,14 @@ impl AdamStates {
 pub struct AdamStatesConv2D {
     pub m: Array4<f32>,
     pub v: Array4<f32>,
+    pub m_bias: Array2<f32>,
+    pub v_bias: Array2<f32>,
+}
+
+#[derive(Debug, Clone, Default)]
+pub struct AdamStatesConv1D {
+    pub m: Array3<f32>,
+    pub v: Array3<f32>,
     pub m_bias: Array2<f32>,
     pub v_bias: Array2<f32>,
 }