+++ changelog.md
@@ -6,6 +6,9 @@ This change log records updates after 2025-3-24.
 Please view [SomeB1oody/RustyML](https://github.com/SomeB1oody/RustyML) for more info.
 
 ## [v0.6.2] - 2025-5-30
+### Added
+- Add `Conv1D` layer implementation and tests
+
 ### Changed
 - Refactor tests to reuse `generate_data` function for pooling layers
 - Rename `OptimizerCacheFEL` to `OptimizerCacheConv2D`